+++ packages/bluesky/lib/bluesky.dart
@@ -133,12 +133,12 @@ export 'package:atproto/atproto.dart'
         ModerationReasonType,
         ReportSubject,
         $ReportSubjectCopyWith,
-        ServersService,
-        IdentitiesService,
-        RepositoriesService,
+        ServerService,
+        IdentityService,
+        RepoService,
         ModerationService,
         SyncService,
-        LabelsService;
+        LabelService;
 export 'package:atproto_core/atproto_core.dart'
     show
         XRPCResponse,
@@ -183,7 +183,7 @@ export 'package:atproto_core/atproto_core.dart'
         refreshSession,
         deleteSession,
         isValidAppPassword;
-export 'package:bluesky/src/services/actors_service.dart';
+export 'package:bluesky/src/services/actor_service.dart';
 export 'package:bluesky/src/bluesky.dart';
 export 'package:bluesky/src/services/services.dart';
 export 'package:bluesky/src/services/entities/actor.dart';
@@ -305,13 +305,13 @@ export 'package:bluesky/src/services/extensions/grouped_notification_extension.d
 export 'package:bluesky/src/services/extensions/at_uri_extension.dart';
 export 'package:bluesky/src/services/extensions/blob_extension.dart';
 export 'package:bluesky/src/services/extensions/strong_ref_extension.dart';
-export 'package:bluesky/src/services/feeds_service.dart';
-export 'package:bluesky/src/services/graphs_service.dart';
+export 'package:bluesky/src/services/feed_service.dart';
+export 'package:bluesky/src/services/graph_service.dart';
 export 'package:bluesky/src/services/constants/notification_reason.dart';
 export 'package:bluesky/src/services/utils/notification_reason_filter.dart'
     show NotificationReasonFilter;
 export 'package:bluesky/src/services/constants/grouped_notification_reason.dart';
-export 'package:bluesky/src/services/notifications_service.dart';
+export 'package:bluesky/src/services/notification_service.dart';
 export 'package:bluesky/src/services/params/generator_param.dart';
 export 'package:bluesky/src/services/params/list_item_param.dart';
 export 'package:bluesky/src/services/params/list_param.dart';