+++ packages/bluesky/lib/src/bluesky.dart
@@ -7,10 +7,10 @@ import 'package:atproto/atproto.dart' as atp;
 import 'package:atproto_core/atproto_core.dart' as core;
 
 // 🌎 Project imports:
-import 'services/actors_service.dart';
-import 'services/feeds_service.dart';
-import 'services/graphs_service.dart';
-import 'services/notifications_service.dart';
+import 'services/actor_service.dart';
+import 'services/feed_service.dart';
+import 'services/graph_service.dart';
+import 'services/notification_service.dart';
 import 'services/services.dart';
 import 'services/unspecced_service.dart';
 
@@ -63,38 +63,89 @@ sealed class Bluesky {
   /// [Bluesky.fromSession], otherwise null.
   core.Session? get session;
 
-  /// Returns the actors service.
-  ActorsService get actors;
+  /// Returns the actor service.
+  /// This service represents `app.bsky.actor.*`.
+  @Deprecated('Use .actor instead. Will be removed')
+  ActorService get actors;
 
-  /// Returns the feeds service.
-  FeedsService get feeds;
+  /// Returns the actor service.
+  /// This service represents `app.bsky.actor.*`.
+  ActorService get actor;
 
-  /// Returns the notifications service.
-  NotificationsService get notifications;
+  /// Returns the feed service.
+  /// This service represents `app.bsky.feed.*`.
+  @Deprecated('Use .feed instead. Will be removed')
+  FeedService get feeds;
 
-  /// Returns the graphs service.
-  GraphsService get graphs;
+  /// Returns the feed service.
+  /// This service represents `app.bsky.feed.*`.
+  FeedService get feed;
+
+  /// Returns the notification service.
+  /// This service represents `app.bsky.notification.*`.
+  @Deprecated('Use .notification instead. Will be removed')
+  NotificationService get notifications;
+
+  /// Returns the notification service.
+  /// This service represents `app.bsky.notification.*`.
+  NotificationService get notification;
+
+  /// Returns the graph service.
+  /// This service represents `app.bsky.graph.*`.
+  @Deprecated('Use .graph instead. Will be removed')
+  GraphService get graphs;
+
+  /// Returns the graph service.
+  /// This service represents `app.bsky.graph.*`.
+  GraphService get graph;
 
   /// Returns the unspecced service.
+  /// This service represents `app.bsky.unspecced.*`.
   UnspeccedService get unspecced;
 
-  /// Returns the servers service.
-  atp.ServersService get servers;
+  /// Returns the server service.
+  /// This service represents `com.atproto.server.*`.
+  @Deprecated('Use .server instead. Will be removed')
+  atp.ServerService get servers;
 
-  /// Returns the identities service.
-  atp.IdentitiesService get identities;
+  /// Returns the server service.
+  /// This service represents `com.atproto.server.*`.
+  atp.ServerService get server;
 
-  /// Returns the repositories service.
-  atp.RepositoriesService get repositories;
+  /// Returns the identity service.
+  /// This service represents `com.atproto.identity.*`.
+  @Deprecated('Use .identity instead. Will be removed')
+  atp.IdentityService get identities;
+
+  /// Returns the identity service.
+  /// This service represents `com.atproto.identity.*`.
+  atp.IdentityService get identity;
+
+  /// Returns the repo service.
+  /// This service represents `com.atproto.repo.*`.
+  @Deprecated('Use .repo instead. Will be removed')
+  atp.RepoService get repositories;
+
+  /// Returns the repo service.
+  /// This service represents `com.atproto.repo.*`.
+  atp.RepoService get repo;
 
   /// Returns the moderation service.
+  /// This service represents `com.atproto.moderation.*`.
   atp.ModerationService get moderation;
 
   /// Returns the sync service.
+  /// This service represents `com.atproto.sync.*`.
   atp.SyncService get sync;
 
-  /// Returns the labels service.
-  atp.LabelsService get labels;
+  /// Returns the label service.
+  /// This service represents `com.atproto.label.*`.
+  @Deprecated('Use .repo label. Will be removed')
+  atp.LabelService get labels;
+
+  /// Returns the label service.
+  /// This service represents `com.atproto.label.*`.
+  atp.LabelService get label;
 }
 
 final class _Bluesky implements Bluesky {
@@ -147,28 +198,49 @@ final class _Bluesky implements Bluesky {
   final core.Session? session;
 
   @override
-  ActorsService get actors => _service.actors;
+  ActorService get actors => _service.actor;
 
   @override
-  FeedsService get feeds => _service.feeds;
+  ActorService get actor => _service.actor;
 
   @override
-  NotificationsService get notifications => _service.notifications;
+  FeedService get feeds => _service.feed;
 
   @override
-  GraphsService get graphs => _service.graphs;
+  FeedService get feed => _service.feed;
+
+  @override
+  NotificationService get notifications => _service.notification;
+
+  @override
+  NotificationService get notification => _service.notification;
+
+  @override
+  GraphService get graphs => _service.graph;
+
+  @override
+  GraphService get graph => _service.graph;
 
   @override
   UnspeccedService get unspecced => _service.unspecced;
 
   @override
-  atp.ServersService get servers => _service.servers;
+  atp.ServerService get servers => _service.server;
 
   @override
-  atp.IdentitiesService get identities => _service.identities;
+  atp.ServerService get server => _service.server;
 
   @override
-  atp.RepositoriesService get repositories => _service.repositories;
+  atp.IdentityService get identities => _service.identity;
+
+  @override
+  atp.IdentityService get identity => _service.identity;
+
+  @override
+  atp.RepoService get repositories => _service.repo;
+
+  @override
+  atp.RepoService get repo => _service.repo;
 
   @override
   atp.ModerationService get moderation => _service.moderation;
@@ -177,5 +249,8 @@ final class _Bluesky implements Bluesky {
   atp.SyncService get sync => _service.sync;
 
   @override
-  atp.LabelsService get labels => _service.labels;
+  atp.LabelService get labels => _service.label;
+
+  @override
+  atp.LabelService get label => _service.label;
 }